changed a character from circle to x placed at the bottom-right
Fax sheet can have noise. It will be appeared as a small dot image.
If the noise will appear at the bottom-right of the fax sheet,
it will be OCRed to recognize the image from circle, triangle and x.
Kocr recognizes the noise as a circle. So it will lead that the
echo back fax will be sent to the sender of the fax. To prevent this
situation, this commit was made. This idea is from Takashi Okumura.