set bind address to localhost for mysqld
Rails currently uses a socket to communicate with mysqld, does not use tcp socket.
The value of bind address is from ohai, and it will be an address which is set
to the current network interface.
If the static address is set to bind address, it will not be good when dhcp is
used because mysqld cannot bind the address.